+++ Magic/src/main/java/com/elmakers/mine/bukkit/economy/VaultCurrency.java
@@ -51,4 +51,8 @@ public class VaultCurrency extends BaseMagicCurrency {
     public boolean isValid() {
         return isValid;
     }
+
+    protected boolean hasDecimals() {
+        return true;
+    }
 }